refactor(clawdstrike-logos): extract verifier tests into tests/ subtree
Move the inline `mod tests` body out of verifier.rs into verifier/tests/,
split across theme files (core, inheritance, witness_synthesis, load_time,
z3) that are `include!`d from tests/mod.rs so they keep sharing one module's
`use super::*` surface and the three fixtures (agent, formula_verifier,
simple_forbidden_path) -- mirroring the api_server precedent. The root now
declares `#[cfg(test)] mod tests;` and is ~70 lines. No test was deleted,
weakened, ignored, reworded, or otherwise altered (verified by a sorted diff
of the original test region against the combined theme files); the two
#[cfg(feature = "z3")] tests are preserved verbatim in z3.rs.
